requirements.txt
youtube-transcript-api == 1.0.3
google-generativeai == 0.8.5
PyPDF2 == 3.0.1
reportlab == 4.4.0
Flask == 3.1.0
flask-cors == 5.0.1
Werkzeug == 3.1.3
mysql-connector-python == 9.3.0
firebase-admin == 6.8.0